requirements.txt
pyrogram==2.0.106
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
httpx[http2]
motor
pillow
psutil
py-tgcalls==2.2.8
ntgcalls>=2.0.6
pykeyboard
pyrofork
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
unidecode
yt-dlp==2025.12.08
py-yt-search
flask
